+++ indexer/src/fastnear_indexer.rs
